+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - feat: adding `NesSelectionList.disabledItems`
  - feat: `NesPressable` is only pressable when has listeners.
  - feat: adding `NesIcons.block`
+ - feat: adding `NesIconButton.badge`
 
 # 0.20.0
  - feat: adding different frames to `NesDialog`s