+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 0.22.0
+ - feat: adding `NesIcons.shovel`.
+
 # 0.21.0
  - Update to Flutter 3.22.0
  - feat: adding `NesDropdownMenu`